Fix #25 (2.1.1)
