+++ restructuredpython.egg-info/PKG-INFO
@@ -1,6 +1,6 @@
 Metadata-Version: 2.4
 Name: restructuredpython
-Version: 2.1.0
+Version: 2.1.1
 Summary: A superset of Python with many new features, including full JS integration, multiline comments, header files, and optional curly brackets around control statements
 Author-email: Rihaan Meher <meherrihaan@gmail.com>
 License: Apache-2.0
