+++ restructuredpython/restructuredpython.py
@@ -1,238 +1,14 @@
 import argparse
 import re
-import sys
 import os
-import warnings
 import tempfile
 from pathlib import Path
-import ctypes
 import sys
 import tomllib as toml
 import fnmatch
-import pkgutil
-import importlib
-import struct
-
-spec = importlib.util.find_spec("restructuredpython")
-if spec and spec.origin:
-    package_dir = os.path.dirname(spec.origin)
-    io_dll = os.path.join(package_dir, "lib", "windows-libs", "io64.dll")
-
-    io32_dll = os.path.join(package_dir, "lib", "windows-lib", "io32.dll")
-
-    io_so = os.path.join(package_dir, "lib", "linux-libs", "io.so")
-
-    io_dylib = os.path.join(package_dir, "lib", "macos-libs", "io.dylib")
-
-if sys.platform == "win32":
-    if (struct.calcsize("P") * 8) == 32:
-        lib = ctypes.WinDLL(io32_dll)
-    else:
-        lib = ctypes.WinDLL(io_dll)
-elif sys.platform == "darwin":
-    lib = ctypes.CDLL(io_dylib)
-else:
-    lib = ctypes.CDLL(io_so)
-
-
-lib.check_file_exists.argtypes = [ctypes.c_char_p]
-lib.check_file_exists.restype = ctypes.c_int
-
-lib.read_file.argtypes = [ctypes.c_char_p]
-lib.read_file.restype = ctypes.c_char_p
-
-lib.write_file.argtypes = [ctypes.c_char_p, ctypes.c_char_p]
-lib.write_file.restype = ctypes.c_int
-
-lib.read_binary_file.argtypes = [
-    ctypes.c_char_p, ctypes.POINTER(
-        ctypes.c_size_t)]
-lib.read_binary_file.restype = ctypes.POINTER(ctypes.c_char)
-
-
-def load_toml_binary(filename):
-    filename = str(filename)
-    size = ctypes.c_size_t()
-    raw_data_ptr = lib.read_binary_file(filename.encode(), ctypes.byref(size))
-
-    if not raw_data_ptr:
-        raise FileNotFoundError(f"Could not read {filename}")
-
-    raw_data = ctypes.string_at(raw_data_ptr, size.value)
-    return toml.loads(raw_data.decode())
-
-
-def read_file_utf8(filename: str) -> str:
-    size = ctypes.c_size_t()
-    filename_bytes = filename.encode('utf-8')
-
-    ptr = lib.read_binary_file(filename_bytes, ctypes.byref(size))
-    if not ptr:
-        raise FileNotFoundError(f"File not found: {filename}")
-
-    raw_bytes = ctypes.string_at(ptr, size.value)
-
-    try:
-        text = raw_bytes.decode('utf-8')
-    except UnicodeDecodeError as e:
-        raise ValueError(f"File is not valid UTF-8: {e}")
-
-    return text
-
-
-token_specification = [
-    ('COMMENT', r'/\*.*?\*/'),
-    ('IF', r'if'),
-    ('FOR', r'for'),
-    ('WHILE', r'while'),
-    ('DEF', r'def'),
-    ('ELIF', r'elif'),
-    ('ELSE', r'else'),
-    ('TRY', r'try'),
-    ('EXCEPT', r'except'),
-    ('CLASS', r'class'),
-    ('WITH', r'with'),
-    ('MATCH', r'match'),
-    ('CASE', r'case'),
-    ('PIPE', r'\|>'),
-    ('IDENT', r'[A-Za-z_][A-Za-z0-9_]*'),
-    ('NUMBER', r'\d+'),
-    ('LBRACE', r'\{'),
-    ('RBRACE', r'\}'),
-    ('LPAREN', r'\('),
-    ('RPAREN', r'\)'),
-    ('EQUALS', r'='),
-    ('SKIP', r'[ \t\n\r]+'),
-    ('MISMATCH', r'.'),
-]
-
-token_regex = '|'.join(
-    f'(?P<{pair[0]}>{pair[1]})' for pair in token_specification)
-
-
-def tokenize(code):
-    """Tokenizes the rePython source code and handles multiline comments."""
-    inside_multiline_comment = False
-    for mo in re.finditer(token_regex, code):
-        kind = mo.lastgroup
-        value = mo.group()
-
-        if kind == 'SKIP':
-            continue
-
-        elif kind == 'COMMENT':
-            if value.startswith('/*') and value.endswith('*/'):
-                comment_lines = value[2:-2].splitlines()
-                for line in comment_lines:
-                    yield 'COMMENT', f"# {line.strip()}"
-                continue
-        elif kind == 'MISMATCH':
-            warnings.warn(
-                f"Unexpected character {
-                    value!r}. Continuing with compilation")  # fmt: skip
-            yield kind, value
-
-        else:
-            yield kind, value
-
-
-def check_syntax(input_lines):
-    for i in range(len(input_lines)):
-        line = input_lines[i].strip()
-
-        if line.startswith(('} else', '} elif')):
-            raise SyntaxError(
-                f"Misplaced '{line}' statement at line {
-                    i + 1}. (REPY-0001)")  # fmt: skip
-        if line.startswith('} except'):
-            raise SyntaxError(
-                f"Misplaced 'except' statement at line {
-                    i + 1}. (REPY-0002)")  # fmt: skip
-        if line.startswith('} def'):
-            raise SyntaxError(
-                f"Misplaced 'def' statement at line {
-                    i + 1}. (REPY-0003)")  # fmt: skip
-        if line.startswith('} class'):
-            raise SyntaxError(
-                f"Misplaced 'class' statement at line {
-                    i + 1}. (REPY-0004)")  # fmt: skip
-        if line.startswith('} case'):
-            raise SyntaxError(
-                f"Misplaced 'case' statement at line {
-                    i + 1}. (REPY-0005)")  # fmt: skip
-
-
-def parse_repython(code):
-    """Parses the rePython code and converts it to valid Python code."""
-    def chain_pipeline(code):
-        parts = [part.strip() for part in code.split('|>')]
-        if len(parts) > 1:
-            def nest(parts):
-                if len(parts) == 1:
-                    return parts[0]
-                else:
-                    return f'{parts[-1]}({nest(parts[:-1])})'
-
-            variable, pipeline = parts[0].split(
-                '=') if '=' in parts[0] else ('', parts[0])
-            variable = variable.strip()
-            pipeline = pipeline.strip()
-
-            if variable:
-                nested_call = nest([pipeline] + parts[1:])
-                return f'{variable} = {nested_call}'
-            else:
-                return nest(parts)
-        return code
-
-    modified_code = []
-    inside_block = False
-    brace_stack = []
-    lines = code.splitlines()
-
-    check_syntax(lines)
-
-    inside_comment_block = False
-
-    for line in lines:
-        processed_line = chain_pipeline(line)
-
-        if inside_comment_block:
-            if processed_line.endswith("*/"):
-                modified_code.append(f"# {processed_line[:-2].strip()}")
-                inside_comment_block = False
-            else:
-                modified_code.append(f"# {processed_line.strip()}")
-        elif processed_line.startswith("/*") and processed_line.endswith("*/"):
-            modified_code.append(f"# {processed_line[2:-2].strip()}")
-        elif processed_line.startswith("/*"):
-            modified_code.append(f"# {processed_line[2:].strip()}")
-            inside_comment_block = True
-        elif processed_line.endswith("*/"):
-            modified_code.append(f"# {processed_line[:-2].strip()}")
-        else:
-            if re.match(
-                r'^\s*(if|for|while|def|try|elif|else|except|class|match|with|case)\s.*\{',
-                    processed_line):
-                modified_code.append(processed_line.split('{')[0] + ':')
-                brace_stack.append('{')
-                inside_block = True
-            elif re.match(r'^\s*\}', processed_line) and inside_block:
-                brace_stack.pop()
-                inside_block = len(brace_stack) > 0
-            elif re.match(r'^\s*match\(', processed_line):
-                modified_code.append(processed_line.split('{')[0] + ':')
-                brace_stack.append('{')
-                inside_block = True
-            elif re.match(r'^\s*case', processed_line):
-                modified_code.append(processed_line.split('{')[0] + ':')
-                brace_stack.append('{')
-                inside_block = True
-            else:
-                modified_code.append(processed_line)
-
-    return '\n'.join(modified_code)
-
+from restructuredpython.parser import parse_repython
+from restructuredpython import cload
+from restructuredpython.cload import *
 
 def compile_header_file(header_filename):
     """Compiles a .cdata file and returns the corresponding Python code."""
